api/requirements.txt
fastapi
uvicorn
python-dotenv
groq
google-generativeai
requests
